GiftCardSummary refactor for CE and EE.
