+++ packages/venia-ui/lib/components/CartPage/PriceSummary/GiftCardSummary.ce.js
@@ -1,0 +1,1 @@
+export default () => null;
